refactor: 객관식 문제 반환 DTO Generic으로 통합 적용
- 정적 팩토리 메소드 1개 감소
