+++ src/main/java/com/server/computer_science/question/common/domain/ChoiceProvider.java
@@ -1,0 +1,7 @@
+package com.server.computer_science.question.common.domain;
+
+import java.util.List;
+
+public interface ChoiceProvider {
+    List<? extends QuestionChoice> getQuestionChoices();
+}
