+++ src/main/java/com/server/computer_science/question/license_question/controller/LicenseQuestionGetViewController.java
@@ -3,7 +3,7 @@ package com.server.computer_science.question.license_question.controller;
 import com.server.computer_science.question.license_question.dto.response.ResponseLicenseSessionDto;
 import com.server.computer_science.question.license_question.service.LicenseQuestionGetService;
 import com.server.computer_science.question.license_question.service.LicenseSessionService;
-import com.server.computer_science.question.major_question.user.dto.response.ResponseClassifiedMultipleQuestionDto;
+import com.server.computer_science.question.common.dto.response.ResponseClassifiedMultipleQuestionDto;
 import io.swagger.annotations.Api;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
