+++ src/main/java/com/server/computer_science/question/major_question/admin/service/implement/BasicAdminMajorQuestionClassifiedGetService.java
@@ -4,7 +4,7 @@ package com.server.computer_science.question.major_question.admin.service.implem
 import com.server.computer_science.question.major_question.common.domain.MajorMultipleChoiceQuestion;
 import com.server.computer_science.question.major_question.common.service.QuestionClassifyByCategoryService;
 import com.server.computer_science.question.major_question.user.dto.request.RequestGetQuestionByCategoryAndLevelDto;
-import com.server.computer_science.question.major_question.user.dto.response.ResponseClassifiedMultipleQuestionDto;
+import com.server.computer_science.question.common.dto.response.ResponseClassifiedMultipleQuestionDto;
 import com.server.computer_science.question.major_question.common.service.MajorQuestionClassifiedGetService;
 import com.server.computer_science.question.major_question.common.service.implement.MajorMultipleChoiceQuestionDBService;
 import lombok.RequiredArgsConstructor;
@@ -20,7 +20,7 @@ public class BasicAdminMajorQuestionClassifiedGetService implements MajorQuestio
     private final QuestionClassifyByCategoryService questionClassifyByCategoryService;
 
     @Override
-    public List<ResponseClassifiedMultipleQuestionDto> getClassifiedNormalQuestions(RequestGetQuestionByCategoryAndLevelDto requestGetQuestionByCategoryAndLevelDto) {
+    public List<ResponseClassifiedMultipleQuestionDto> getClassifiedMajorMultipleChoiceQuestions(RequestGetQuestionByCategoryAndLevelDto requestGetQuestionByCategoryAndLevelDto) {
         List<MajorMultipleChoiceQuestion> majorMultipleChoiceQuestions = majorMultipleChoiceQuestionDBService.getFetchChoicesByCategoriesAndLevels(requestGetQuestionByCategoryAndLevelDto.getQuestionCategories(), requestGetQuestionByCategoryAndLevelDto.getQuestionLevels());
         return questionClassifyByCategoryService.classifyQuestionByCategoryOrdered(majorMultipleChoiceQuestions)
                 .entrySet().stream()
