+++ src/main/java/com/server/computer_science/question/major_question/common/service/MajorQuestionClassifiedGetService.java
@@ -2,7 +2,7 @@ package com.server.computer_science.question.major_question.common.service;
 
 
 import com.server.computer_science.question.major_question.user.dto.request.RequestGetQuestionByCategoryAndLevelDto;
-import com.server.computer_science.question.major_question.user.dto.response.ResponseClassifiedMultipleQuestionDto;
+import com.server.computer_science.question.common.dto.response.ResponseClassifiedMultipleQuestionDto;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -12,7 +12,7 @@ public interface MajorQuestionClassifiedGetService {
     /**
      * 분류별로 나누어 문제를 반환
      */
-    public List<ResponseClassifiedMultipleQuestionDto> getClassifiedNormalQuestions(RequestGetQuestionByCategoryAndLevelDto requestGetQuestionByCategoryAndLevelDto);
+    public List<ResponseClassifiedMultipleQuestionDto> getClassifiedMajorMultipleChoiceQuestions(RequestGetQuestionByCategoryAndLevelDto requestGetQuestionByCategoryAndLevelDto);
     public List<ResponseClassifiedMultipleQuestionDto> getClassifiedShortAnsweredMajorQuestions(RequestGetQuestionByCategoryAndLevelDto requestGetQuestionByCategoryAndLevelDto);
     public List<ResponseClassifiedMultipleQuestionDto> getClassifiedAllMajorQuestions();
 }
